train val split in engine.cpp + normalization implementation

+++ data_utils.cpp
@@ -16,8 +16,8 @@ void train_val_split_indices_deterministic(
     size_t array_size, 
     double train_ratio, 
     unsigned fixed_seed,
-    std::vector<size_t>& train_indices_out,
-    std::vector<size_t>& val_indices_out) 
+    std::vector<uint16_t>& train_indices_out,
+    std::vector<uint16_t>& val_indices_out) 
 {
     // Clear previous contents
     train_indices_out.clear();
